chore: extract Package Validator [OSM-2240]
In preparation for feeding cached test result into the validation step,
extracting the validator to its own class.

This commit does not introduce any behavioural changes, only lays
the ground for the next iteration. In a follow-up step, we will allow
skipping of the test so that gatekeeping is based on cached results.