chore: introduce artifact model [OSM-2240]
In order to let the validation run against cached artifact metadata
rather than a full test result, introducing a model representing vuln
summary.

For the time being, the model is always created from a test result but
in a follow-up PR we'll also support deriving it from cached data. This
way we'll be able to avoid making an unnecessary Snyk Test request when
there is fresh vuln information already.