feat: switch first lifecycle hook to afterCreate
Switching the hook which gates artifact downloads from `afterRemoteDownload`
to `afterCreate`.

For some ecosystems, e.g. Python and Cocoapods, properties would fail to
write during the `afterRemoteDownload`. This would manifest with the
following error message in Artifactory logs:

```
Cannot add properties for <repository path> Item not found.
```

The `afterCreate` hook is also attached to initial downloads of
artifacts, and properties are successfully created in this context
across all ecosystems supported by this plugin.